add staking query and remove old query
